fix instructions in F and D extensions

+++ src/register.rs
@@ -25,21 +25,21 @@ impl Index<usize> for Registers {
 }
 
 pub(crate) struct FRegisters {
-  regs: [f64; 32],
+  regs: [u64; 32],
 }
 
 impl FRegisters {
   pub(crate) fn new() -> FRegisters {
-    FRegisters { regs: [0f64; 32] }
+    FRegisters { regs: [0; 32] }
   }
 
-  pub(crate) fn set(&mut self, index: usize, value: f64) {
+  pub(crate) fn set(&mut self, index: usize, value: u64) {
     self.regs[index] = value;
   }
 }
 
 impl Index<usize> for FRegisters {
-  type Output = f64;
+  type Output = u64;
 
   fn index(&self, index: usize) -> &Self::Output {
     &self.regs[index]
